Get most hipsparse tests passing with cuda 11.1 backend (#169)
* dnvec tests working with cuda 11 backend

* spvec tests working with cuda 11 backend

* spmat tests working with cuda 11 backend

* spvv tests working with cuda 11 backend

* spmm tests working with cuda 11 backend

* spmv tests working with cuda 11 backend

* fixing tests for cuda backend

* fixing csx2dense and dense2csx

* fixes

* fixing prune methods for cuda backend

* turn off tests for csrsm2 when using cuda backend

* removing unneeded include

* only run cuda tests when using cuda 11.1 or greater

* only run cuda tests when using cuda 11.1 or greater

* only run cuda tests when using cuda 11.1 or greater

* get csr2csr_compress tests passing for cuda backend

* all tests passing with cuda 11.1 or greater backend

* fix indenting

* fixing indenting

* changing tests for spmm_coo to pass cusparse


[ROCm/hipSPARSE commit: 65dd18e1c0cd025c24af3f1c4126c35350d70a95]

+++ clients/tests/test_spmm_coo.cpp
@@ -39,12 +39,12 @@ TEST(spmm_coo, spmm_coo_i32_float)
 
 TEST(spmm_coo, spmm_coo_i64_double)
 {
-    hipsparseStatus_t status = testing_spmm_coo<int64_t, double>();
+    hipsparseStatus_t status = testing_spmm_coo<int32_t, double>();
     EXPECT_EQ(status, HIPSPARSE_STATUS_SUCCESS);
 }
 
 TEST(spmm_coo, spmm_coo_i64_hipComplex)
 {
-    hipsparseStatus_t status = testing_spmm_coo<int64_t, hipComplex>();
+    hipsparseStatus_t status = testing_spmm_coo<int32_t, hipComplex>();
     EXPECT_EQ(status, HIPSPARSE_STATUS_SUCCESS);
 }
